refactor: naming, code quality
