+++ src/components/TranscriptionInterface.tsx
@@ -14,7 +14,7 @@ import {SpeechTranscriber} from "../utils/speech-transcriber.ts";
 import Controls from "./Controls.tsx";
 import TranscriptBox from "./TranscriptBox.tsx";
 import {translate} from "../utils/translation.ts";
-import { detectKeyPhrases } from "../utils/keyphrase-extraction.ts";
+import { comprehend } from "../utils/comprehend.ts";
 
 function TranscriptionInterface() {
     const speechTranscriber = useRef<SpeechTranscriber | null>(null);
@@ -73,7 +73,7 @@ function TranscriptionInterface() {
         }
 
         const translated = await translate(transcriptPartText, language.translateCode, otherLanguage.translateCode);
-        const keyphrases = await detectKeyPhrases(translated, otherLanguage.translateCode);
+        const keyphrases = await comprehend('comprehendKeyPhrases', translated, otherLanguage.translateCode);
         console.log(keyphrases);
