+++ src/utils/comprehend.ts
@@ -1,13 +1,13 @@
-import Auth from '@aws-amplify/auth';
+import { Amplify, fetchAuthSession } from '@aws-amplify/core';
 import {Lambda} from '@aws-sdk/client-lambda'
 
 /**
- * Detect sentiment in the given text using Lambda function
+ * Detect key phrases in the given text using Lambda function
  * @param text - Input text to analyze
  * @param languageCode - Language code for the text (default: 'en')
  * @returns Promise resolving to an array of key phrases
  */
-export const detectSentimentEN = async (text: string, languageCode: string = 'en') => {
+export const comprehend = async (fName: string ,text: string, languageCode: string = 'en') => {
   // Validate input
   if (!text.trim()) {
     throw new Error('Text cannot be empty');
@@ -15,12 +15,26 @@ export const detectSentimentEN = async (text: string, languageCode: string = 'en
 
   try {
     // Get current credentials from Amplify Auth
-    const credentials = (await Auth.fetchAuthSession()).credentials;
+    const region = Amplify.getConfig()
+        .Predictions!
+        .convert!
+        .transcription!
+        .region!;
+    const authSession = await fetchAuthSession();
+    var credentials = authSession.credentials!;
+    const config = {
+        region: region,
+        credentials: {
+            accessKeyId: credentials.accessKeyId,
+            secretAccessKey: credentials.secretAccessKey,
+            sessionToken: credentials.sessionToken,
+        },
+    };
     
     // Initialize Lambda client with credentials
     const lambda = new Lambda({
       region: 'eu-west-2', 
-      credentials: credentials
+      credentials: config.credentials
     });
 
     // Prepare the payload
@@ -31,28 +45,25 @@ export const detectSentimentEN = async (text: string, languageCode: string = 'en
 
     // Invoke Lambda function
     const response = await lambda.invoke({
-      FunctionName: 'comprehendSentimentDetection', // Replace with your actual function name
+      FunctionName: fName, 
       Payload: JSON.stringify({ body: JSON.stringify(payload) })
     });
 
-    // Parse the response
     if (!response.Payload) {
       throw new Error('Empty response from Lambda function');
     }
-
-    const responseBody = JSON.parse('');
     
     // If the Lambda returned an error
-    if (responseBody.statusCode !== 200) {
-      throw new Error(responseBody.body ? JSON.parse(responseBody.body).error : 'Lambda function error');
+    if (response.StatusCode !== 200) {
+      throw new Error('Lambda function error');
     }
 
     // Extract and return the key phrases
-    const parsedBody = JSON.parse(responseBody.body);
-    return parsedBody || [];
+    const parsedBody = new TextDecoder().decode(response.Payload);
+    return JSON.parse(parsedBody).body;
+
   } catch (error) {
     console.error('Key Phrases Detection Error:', error);
     throw error;
   }
 };
-
